+++ packages/store/ts/config/v2/compat.ts
@@ -13,10 +13,6 @@ export type storeToV1<store> = store extends Store
           filePath: string;
         };
       };
-      storeImportPath: store["codegen"]["storeImportPath"];
-      userTypesFilename: store["codegen"]["userTypesFilename"];
-      codegenDirectory: store["codegen"]["outputDirectory"];
-      codegenIndexFilename: store["codegen"]["indexFilename"];
       tables: {
         [key in keyof store["tables"] as store["tables"][key]["name"]]: tableToV1<store["tables"][key]>;
       };
@@ -27,7 +23,7 @@ export type storeToV1<store> = store extends Store
 type schemaToV1<schema extends Schema> = { [key in keyof schema]: schema[key]["internalType"] };
 
 export type tableToV1<table extends Table> = {
-  directory: table["codegen"]["outputDirectory"];
+  directory: string;
   dataStruct: table["codegen"]["dataStruct"];
   tableIdArgument: table["codegen"]["tableIdArgument"];
   storeArgument: table["codegen"]["storeArgument"];
@@ -65,10 +61,6 @@ export function storeToV1<store>(store: conform<store, Store>): storeToV1<store>
     namespace: store.namespace,
     enums: store.enums,
     userTypes: resolvedUserTypes,
-    storeImportPath: store.codegen.storeImportPath,
-    userTypesFilename: store.codegen.userTypesFilename,
-    codegenDirectory: store.codegen.outputDirectory,
-    codegenIndexFilename: store.codegen.indexFilename,
     tables: resolvedTables,
     v2: store,
   } as never;