+++ packages/world/ts/config/v2/defaults.ts
@@ -1,16 +1,26 @@
+import { CodegenInput, DeployInput, ModuleInput, SystemInput, WorldInput } from "./input";
+
 export const SYSTEM_DEFAULTS = {
   registerFunctionSelectors: true,
   openAccess: true,
-  accessList: [] as string[],
-} as const;
+  accessList: [],
+} as const satisfies SystemInput;
 
 export type SYSTEM_DEFAULTS = typeof SYSTEM_DEFAULTS;
+
+export const MODULE_DEFAULTS = {
+  root: false,
+  args: [],
+  artifactPath: undefined,
+} as const satisfies Pick<ModuleInput, "root" | "args" | "artifactPath">;
+
+export type MODULE_DEFAULTS = typeof MODULE_DEFAULTS;
 
 export const CODEGEN_DEFAULTS = {
   worldInterfaceName: "IWorld",
   worldgenDirectory: "world",
   worldImportPath: "@latticexyz/world/src/",
-} as const;
+} as const satisfies CodegenInput;
 
 export type CODEGEN_DEFAULTS = typeof CODEGEN_DEFAULTS;
 
@@ -20,17 +30,17 @@ export const DEPLOY_DEFAULTS = {
   deploysDirectory: "./deploys",
   worldsFile: "./worlds.json",
   upgradeableWorldImplementation: false,
-} as const;
+} as const satisfies DeployInput;
 
 export type DEPLOY_DEFAULTS = typeof DEPLOY_DEFAULTS;
 
 export const CONFIG_DEFAULTS = {
   systems: {},
   tables: {},
-  excludeSystems: [] as string[],
+  excludeSystems: [],
   modules: [],
   codegen: CODEGEN_DEFAULTS,
   deploy: DEPLOY_DEFAULTS,
-} as const;
+} as const satisfies WorldInput;
 
 export type CONFIG_DEFAULTS = typeof CONFIG_DEFAULTS;